---------- My portfolio page ----------
I set the html skeleton (the header and the footer)
The main content body of the page is in the main which contains acceuil, Résumé, travail, contactez (home, resume, work, contact).

I use "data-toggle" and "data-target" to show the correspond content in the main page.

I also use bootStrap and font awesome for my css style and icon selection.